refactor: rename NativeBridge to NativeRNBridge and update imports
